Mod.fun search in autocomplete - first draft
